renamed extend and PowerFailed
