+++ TFT/src/User/Menu/Print.c
@@ -352,7 +352,7 @@ void menuPrint(void)
         list_mode = infoSettings.file_listmode;  // follow list mode setting in TFT sd card
         infoFile.source = TFT_SD;
         infoMenu.menu[++infoMenu.cur] = menuPrintFromSource;
-        infoMenu.menu[++infoMenu.cur] = menuPowerOff;
+        infoMenu.menu[++infoMenu.cur] = menuPrintRestore;
         goto selectEnd;
 
       #ifdef U_DISK_SUPPORT
@@ -360,7 +360,7 @@ void menuPrint(void)
           list_mode = infoSettings.file_listmode;  // follow list mode setting in usb disk
           infoFile.source = TFT_UDISK;
           infoMenu.menu[++infoMenu.cur] = menuPrintFromSource;
-          infoMenu.menu[++infoMenu.cur] = menuPowerOff;
+          infoMenu.menu[++infoMenu.cur] = menuPrintRestore;
           goto selectEnd;
         case KEY_ICON_2:
       #else
