+++ TFT/src/User/Menu/PrintRestore.h
@@ -1,0 +1,14 @@
+#ifndef _PRINT_RESTORE_H_
+#define _PRINT_RESTORE_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void menuPrintRestore(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
