+++ TFT/src/User/includes.h
@@ -69,7 +69,6 @@
 #include "config.h"
 #include "coordinate.h"
 #include "debug.h"
-#include "extend.h"
 #include "FanControl.h"
 #include "flashStore.h"
 #include "HomeOffsetControl.h"
@@ -84,12 +83,14 @@
 #include "ModeSwitching.h"
 #include "Notification.h"
 #include "parseACK.h"
+#include "PowerFailed.h"
 #include "Printing.h"
 #include "ProbeHeightControl.h"
 #include "ProbeOffsetControl.h"
 #include "ScreenShot.h"
 #include "serialConnection.h"
 #include "Settings.h"
+#include "SmartFeatures.h"
 #include "SpeedControl.h"
 #include "Temperature.h"
 #include "Touch_Encoder.h"
@@ -126,10 +127,10 @@
 #include "PersistentInfo.h"
 #include "Pid.h"
 #include "Popup.h"
-#include "PowerFailed.h"
 #include "PreheatMenu.h"
 #include "Print.h"
 #include "PrintingMenu.h"
+#include "PrintRestore.h"
 #include "RRFMacros.h"
 #include "ScreenSettings.h"
 #include "SelectMode.h"
